docs: clarify flick timing tooltip

+++ mm/2s2h/Enhancements/Player/ModernZTargeting.cpp
@@ -1,4 +1,4 @@
-#include <cstdlib>
+#include <cmath>
 #include <libultraship/bridge/consolevariablebridge.h>
 #include "2s2h/GameInteractor/GameInteractor.h"
 #include "2s2h/ShipInit.hpp"
@@ -126,7 +126,7 @@ static Actor* FindTargetActor(PlayState* play, Player* player, s32 switchDirecti
             s16 actorYaw = Math_Vec3f_Yaw(&player->actor.world.pos, &actor->focus.pos);
             s16 relYaw = (s16)(actorYaw - cameraYaw);
             s16 yawDiff = (s16)(relYaw - currentRel);
-            s32 absYawDiff = std::abs(yawDiff);
+            s32 absYawDiff = std::abs(static_cast<s32>(yawDiff));
 
             if (switchDirection == SWITCH_DIRECTION_NEAREST) {
                 if ((absYawDiff != 0) && (!foundDirect || (absYawDiff < bestAbsDiff))) {
